ir: Speedup Schedule through QueueStateful
