+++ devito/ir/clusters/queue.py
@@ -1,6 +1,10 @@
+from collections import defaultdict
 from itertools import groupby
 
-__all__ = ['Queue']
+from devito.ir.support import Scope
+from devito.tools import as_tuple, flatten
+
+__all__ = ['Queue', 'QueueStateful']
 
 
 class Queue(object):
@@ -71,3 +75,45 @@ class Queue(object):
                 processed.extend(self._process_fatd(_clusters, level + 1))
 
         return processed
+
+
+class QueueStateful(Queue):
+
+    """
+    A Queue carrying along some state. This is useful when one wants to avoid
+    expensive re-computations of information.
+    """
+
+    class State(object):
+
+        def __init__(self):
+            self.properties = {}
+            self.scopes = {}
+
+    def __init__(self, state=None):
+        super(QueueStateful, self).__init__()
+        self.state = state or QueueStateful.State()
+
+    def _fetch_scope(self, clusters):
+        exprs = flatten(c.exprs for c in as_tuple(clusters))
+        key = tuple(exprs)
+        if key not in self.state.scopes:
+            self.state.scopes[key] = Scope(exprs)
+        return self.state.scopes[key]
+
+    def _fetch_properties(self, clusters, prefix):
+        # If the situation is:
+        #
+        # t
+        #   x0
+        #     <some clusters>
+        #   x1
+        #     <some other clusters>
+        #
+        # then retain only the "common" properties, that is those along `t`
+        properties = defaultdict(set)
+        for c in clusters:
+            v = self.state.properties.get(c, {})
+            for i in prefix:
+                properties[i.dim].update(v.get(i.dim, set()))
+        return properties
